Improve Mechanics in general (moving, splitting, etc)

+++ src/main/java/model/Cell.java
@@ -16,8 +16,8 @@ public abstract class Cell {
   private int radius;
   private int mass;
 
-  private int speedX;
-  private int speedY;
+  private float speedX;
+  private float speedY;
 
   public Cell(int x, int y, int mass) {
     this.x = x;
@@ -42,19 +42,19 @@ public abstract class Cell {
     this.y = y;
   }
 
-  public int getSpeedX() {
+  public float getSpeedX() {
     return speedX;
   }
 
-  public void setSpeedX(int speedX) {
+  public void setSpeedX(float speedX) {
     this.speedX = speedX;
   }
 
-  public int getSpeedY() {
+  public float getSpeedY() {
     return speedY;
   }
 
-  public void setSpeedY(int speedY) {
+  public void setSpeedY(float speedY) {
     this.speedY = speedY;
   }
